Rename file SouncloudBlockComponent
